requirements.txt
fastapi>=0.110
uvicorn>=0.27
pydantic>=2.6
PyYAML>=6.0
